Maintain ordering of alternatives in environment
The environment file was populating a regular (unordered) Python
dictionary with alternatives, and then the tests were asserting
on the order of these alternatives via the `get_alts()` function.
As a result, these tests were fragile: the order of the list could
change at any time, and if it did, these tests would fail. In fact,
the tests were asserting on an ordering that did not match the ordering
in the .lektorproject file.

With this change, the ordering of the alternatives is maintained
in the environment. The output of the `get_alts()` is consistent, and
can be safely asserted on.